### 2024-09-14 #7224 @dfm
In the meantime, since halotools has previously been published can you provide more details about exactly which parts of the code should be considered for this review? It can be useful to include links to the specific git diffs, and any other context for consideration. Thanks!
