### 2024-09-21 #7224 @dfm
@nvanalfen — Thanks for the info. It's important that the reviewers know exactly which parts of the code they should be reviewing because they will need to complete the [JOSS review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html), so you'll probably need to help them throughout the process to figure out how to review the specific contributions related to this submission. You summary with commit links is a great place to start - please share that again in the main review once it starts!

Regarding the license: don't worry about that check. I also found the file, and as long as the reviewers can find it too, we're not concerned about the automatic check.

For the DOIs, please add the recommended DOIs if they do correspond to the correct articles and removing the `doi.org` prefix is sufficient for the last one. For any references that truly don't have DOIs it's fine to ignore the warnings posted by the bot - we do get false positives!
