### 2024-09-23 e436c30
Fix missing and invalid DOI
