+++ paper.bib
@@ -318,6 +318,7 @@ adsnote = {Provided by the SAO/NASA Astrophysics Data System}
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Equatorial Distributions on a Sphere},
  urldate = {2023-03-20},
+ doi = {10.2307/2333824},
  volume = {52},
  year = {1965}
 }
@@ -578,7 +579,7 @@ volume = {114},
 pages = {256-269},
 year = {2013},
 issn = {0047-259X},
-doi = {https://doi.org/10.1016/j.jmva.2012.08.010},
+doi = {10.1016/j.jmva.2012.08.010},
 url = {https://www.sciencedirect.com/science/article/pii/S0047259X12002084},
 author = {Suvrit Sra and Dmitrii Karp},
 keywords = {Watson distribution, Kummer function, Confluent hypergeometric, Directional statistics, Diametrical clustering, Special function, Hypergeometric identities},
@@ -1284,6 +1285,7 @@ archivePrefix = {arXiv},
          year = 2017,
         month = jan,
        adsurl = {https://ui.adsabs.harvard.edu/abs/2017PhDT.......522S},
+       doi = {10.2172/1420403},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
