### 2024-10-02 #7224 @ivastar
@nvanalfen is the plan to eventually merge the contributions on this fork with the main `halotools` repo? Will this be done at publication?
