### 2024-10-02 #7224 @nvanalfen
@ivastar The contributions have been merged into the main repo and published on PyPI.

Initially, I had forked, made my contributions, then we merged a PR into the main `halotools` repo. When we finished up this paper, Andrew published the new version and I updated my fork to match the current state of the main repo (some other changes had been made since those reflected in this work). That way, the paper branch matched the current state of things.

Also, I'm gathering names of potential reviewers and I'll send an email with those. Thanks!
