### 2024-11-04 #7421 @editorialbot
[REVIEW]: Halotools: A New Release Adding Intrinsic Alignments to Halo Based Methods
**Submitting author:** @nvanalfen (<a href="http://orcid.org/0000-0003-0049-2861">Nicholas Van Alfen</a>)
**Repository:** https://github.com/nvanalfen/halotools
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.9.3
