### 2024-11-04 #7421 @aphearin
Happy to refactor @ivastar - any templates I should look at?
