### 2024-11-05 #7421 @ivastar
> Happy to refactor @ivastar - any templates I should look at?

https://github.com/astropy/halotools/pull/1105 should do it. And then @nvanalfen can pull to his fork to bring the changes over.
