### 2024-11-21 nvanalfen/halotools#9 @cmlamman
Reproducing first IA tutorial
I'm running into issues reproducing hod_modeling_tutorial_with_alignments.ipynb.
I've been able to download and use the example halo catalog (simname='bolplanck', halo_finder='rockstar', redshift=0, version_name='halotools_v0p4'), but when running mask_bad_halocat(halocat) or model_instance.populate_mock(halocat, seed=seed), I get a few errors:
