### 2024-11-25 nvanalfen/halotools#9 @nvanalfen
@cmlamman Interesting. I have two initial idea for what might be causing this. Either the cached catalog hasn't been downloaded, or I've set up something in my environment that's important but have forgotten to document.

For the first point: after first installing halotools, you need to download a cached halo catalog. You only need to do this once (for any cached catalog you want to use). I'll outline that process here, but if this doesn't solve it, it might be that I've set something up in my environment that I haven't documented (in which case I'll very much want to figure out what that is so I can document it).

Have you been able to run any of the non-IA tutorial notebooks?

To download a cached halocat, run the following code:
> from halotools.sim_manager import DownloadManager
dman = DownloadManager()
dman.download_processed_halo_table('bolplanck', 'rockstar', 0)

The full documentation is here if you need a deeper look.
https://halotools.readthedocs.io/en/latest/quickstart_and_tutorials/quickstart_guides/working_with_halotools_provided_catalogs.html
If this doesn't work, please let me know and I'll track down the issue.
