### 2024-11-25 nvanalfen/halotools#9 @cmlamman
Yes, I downloaded the cached halo catalog (simname='bolplanck', halo_finder='rockstar', redshift=0, version_name='halotools_v0p4') that you used in the tutorial and was able to read it into the notebook. I was also able to run other tutorials.
