### 2024-11-25 nvanalfen/halotools#9 @nvanalfen
I made a fresh virtual environment, but still couldn't recreate this error. Can you tell me the following:

- Your version number of halotools (astropy, scipy, and numpy too just in case).
- It looks like the error message you sent was truncated. Other than the `AttributeError: 'CachedHaloCatalog' object has no attribute '_halo_table'` message, can you paste the full message, or the last couple blocks of the untruncated message?
- And this this is happening on the third cell on the `mask_bad_halocat(halocat)` line?
- Anything else you think could help me recreate this error.

Of the two errors I do see in the trace you posted:
- I have seen `AttributeError: 'CachedHaloCatalog' object has no attribute '_halo_table'` before when the cached halocat didn't properly initialize.
- The error `ValueError: Input array y must be a 1d sequence of unique integers` occurs when crossmatch is called with an array with duplicate values or non integers. The fact that we see this error on a notebook where you haven't made any changes (especially if it has been triggered by mask_bad_halocat) does feel odd to me. I hope the full trace can clear it up (if there's more to that error trace)
