### 2024-12-02 nvanalfen/halotools#9 @nvanalfen
@cmlamman Ok, I'm still looking into this but I wanted to give you an (unfortunately not very helpful) update on what I've seen so far. The versions of the packages you're using are the exact same as what I'm running, so no issues there. It ran when I made a fresh virtual environment, but to make sure, I tried on a fresh computer.

On this fresh computer, there was no halotools installation (now or ever), so I installed halotools and tried to run the notebook (`nothing beyond pip3 install halotools` in this environment, as well as downgrading the astropy version to what's given above as there seem to be issues on later versions). I expected (and got) the error that I hadn't yet downloaded a cached halocat. So I ran the command you gave earlier, namely `python3 download_additional_halocat.py bolplanck rockstar halotools_v0p4 0` to download the proper halocat. Then I restarted the notebook kernel and ran the whole thing, but it went smoothly and I wasn't able to recreate the errors you're getting.

You could try uninstalling halotools, then reinstalling (or even simply trying to freshly download the halocat), but let's check a few more things before that.
What's your OS?
What version of python are you running?
Do you have cython installed and what version?
Do you have gcc installed, and what version?

For reference, here's what I'm running (using the fresh computer specs):
Ubuntu 24.04
Python 3.12.3
Cython 3.0.11
gcc 13.2.0

cython and gcc may not be related to the problem at all (it doesn't sound like they should be), but I have had issues related to errors compiling various pyx files and the like. Even in that case, I believe anything higher than gcc 4 worked.

I'm sorry this has been such a struggle. I really want to figure out what's going on so I can document this problem better.
