### 2024-12-02 #7421 @cmlamman
Thanks for your patience @nvanalfen on the above issue - I will reinstall everything in a fresh environment as soon as I get a chance.
What timeline is acceptable to finish the JOSS review? ( @ivastar )
I will be traveling the next two weeks so I will have very limited time until December 16.
