### 2024-12-18 nvanalfen/halotools#9 @matroxel
FWIW, I was able to install things and run the IA tutorial notebooks without any issues.

However, I did notice another minor thing to fix, which is that installing the master branch failed when doing it into the suggested virtualenv installation setup, which installs a python version <3.10, while the master branch expects python>=3.10.
