### 2024-12-18 #7421 @matroxel
I'm happy to approve - I suggested a very minor documentation fix in the open issue related to required python version for the environment, which should be addressed. I did not run into the issues @cmlamman did with running the suggested examples.
