### 2024-12-18 nvanalfen/halotools#9 @nvanalfen
@matroxel this is in reference to the listen Python dependency at https://halotools.readthedocs.io/en/latest/install.html correct? Thank you for pointing that out. That's an important point I had noticed but hadn't realized I didn't update documentation for. I will get on fixing that. Did you see it listed incorrectly anywhere else?
