### 2024-12-18 nvanalfen/halotools#9 @matroxel
Also here: https://github.com/nvanalfen/halotools/blob/master/docs/installing_halotools_with_virtualenv.rst
