### 2024-12-18 nvanalfen/halotools#9 @nvanalfen
@matroxel Thanks! I've fixed those and submitted a PR to the Astropy master branch. Once that's merged, I'll update my paper branch to reflect that just for consistency.
