### 2024-12-18 nvanalfen/halotools#9 @nvanalfen
@matroxel just an update, I changed the required version in the documentation to 3.11 or later as halotools is planning to drop support for 3.10 soon. 3.10 still works for the version you have now (and in this paper), but we're updating the installation docs to suggest >=3.11 now. Thanks for pointing this out!
