### 2024-12-22 nvanalfen/halotools#9 @cmlamman
@nvanalfen I spent more time trying various instillation methods and the issue is with my OS (Windows). I had no problems with installation and the tutorials when I tried it on Linux.
It seems they are known issues with halotools (https://github.com/astropy/halotools/issues/1055, https://github.com/astropy/halotools/issues/1071), not your contribution specifically so I added relevant comments to those existing issues and will finish up the review as is on my linux machine.
