### 2024-12-22 #7421 @cmlamman
@ivastar I've completed my review!
The minor instillation issues (https://github.com/nvanalfen/halotools/issues/9) are more relevant to the full halotools library and did not prevent me from testing @nvanalfen 's additions.
