+++ paper.bib
@@ -11,255 +11,6 @@
 	day=4,
 }
 
-@preamble{ " \newcommand{\noop}[1]{} " }
-
-@ARTICLE{2013ApJ...762..109B,
-       author = {{Behroozi}, Peter S. and {Wechsler}, Risa H. and {Wu}, Hao-Yi},
-        title = "{The ROCKSTAR Phase-space Temporal Halo Finder and the Velocity Offsets of Cluster Cores}",
-      journal = {\apj},
-     keywords = {dark matter, methods: numerical, Astrophysics - Cosmology and Extragalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = "2013",
-        month = "Jan",
-       volume = {762},
-       number = {2},
-          eid = {109},
-        pages = {109},
-          doi = {10.1088/0004-637X/762/2/109},
-archivePrefix = {arXiv},
-       eprint = {1110.4372},
- primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ApJ...762..109B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{mead_etal15_hmcode,
-       author = {{Mead}, A.~J. and {Peacock}, J.~A. and {Heymans}, C. and {Joudaki}, S. and {Heavens}, A.~F.},
-        title = "{An accurate halo model for fitting non-linear cosmological power spectra and baryonic feedback models}",
-      journal = {\mnras},
-     keywords = {gravitational lensing: weak, cosmology: theory, dark energy, large-scale structure of Universe, Astrophysics - Cosmology and Nongalactic Astrophysics},
-         year = 2015,
-        month = dec,
-       volume = {454},
-       number = {2},
-        pages = {1958-1975},
-          doi = {10.1093/mnras/stv2036},
-archivePrefix = {arXiv},
-       eprint = {1505.07833},
- primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2015MNRAS.454.1958M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{chisari_etal19_ccl,
-       author = {{Chisari}, Nora Elisa and {Alonso}, David and {Krause}, Elisabeth and {Leonard}, C. Danielle and {Bull}, Philip and {Neveu}, J{\'e}r{\'e}my and {Villarreal}, Antonia Sierra and {Singh}, Sukhdeep and {McClintock}, Thomas and {Ellison}, John and {Du}, Zilong and {Zuntz}, Joe and {Mead}, Alexander and {Joudaki}, Shahab and {Lorenz}, Christiane S. and {Tr{\"o}ster}, Tilman and {Sanchez}, Javier and {Lanusse}, Francois and {Ishak}, Mustapha and {Hlozek}, Ren{\'e}e and {Blazek}, Jonathan and {Campagne}, Jean-Eric and {Almoubayyed}, Husni and {Eifler}, Tim and {Kirby}, Matthew and {Kirkby}, David and {Plaszczynski}, St{\'e}phane and {Slosar}, An{\v{z}}e and {Vrastil}, Michal and {Wagoner}, Erika L. and {LSST Dark Energy Science Collaboration}},
-        title = "{Core Cosmology Library: Precision Cosmological Predictions for LSST}",
-      journal = {\apjs},
-     keywords = {cosmology: theory, dark energy, large-scale structure of universe, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2019,
-        month = may,
-       volume = {242},
-       number = {1},
-          eid = {2},
-        pages = {2},
[…]
