### 2025-01-21 4b42f66
Minor editorial changes suggested by editor
