+++ paper.md
@@ -35,36 +35,36 @@ bibliography: paper.bib
 ---
 
 # Summary
-`Halotools`, initially published in 2017, is a Python package for cosmology and astrophysics designed to generate mock universes using existing catalogs of dark matter halos [@Hearin_2017]. A theoretical basis of the library is the so-called halo model, which describes the matter distribution of dark matter as gravitationally self-bound clouds of dark matter particles that we call halos. `Halotools` was designed to take an underlying catalog of dark matter halos and populate them with galaxies using subhalo abundance, or halo occupation distribution (HOD) models, creating catalogs of simulated galaxies for use in research. This release (v0.9) adds functionality to align galaxies, injecting what are known as intrinsic alignments (IA) into these catalogs. As such, these simulated galaxy catalogs can now be created with realistically complex correlations between galaxies, mimicking some effects seen in more expensive hydrodynamic simulations.
+`Halotools`, originally published in 2017, is a Python package for cosmology and astrophysics designed to generate mock universes using existing catalogs of dark matter halos [@Hearin_2017]. A theoretical basis of the library is the so-called halo model, that describes the matter distribution of dark matter as gravitationally self-bound clouds of dark matter particles that we call halos. `Halotools` is designed to take an underlying catalog of dark matter halos and populate them with galaxies using subhalo abundance, or halo occupation distribution (HOD) models, creating catalogs of simulated galaxies for use in research. This release (v0.9) adds functionality to align galaxies, injecting what are known as intrinsic alignments (IA) into these catalogs. As a result, these simulated galaxy catalogs can now be created with realistically complex correlations between galaxies, mimicking some effects seen in more expensive hydrodynamic simulations.
 
 # Statement of Need
-Following the halo model, galaxies form within dark matter halos, and the intrinsic shapes and orientations of these galaxies may be related to those of the host halo and with the large-scale structure of the universe (e.g. the local gravitational tidal field), an effect known as intrinsic alignments (IA) [see, e.g., @Hirata_2004; @Blazek_2019]. The observed shapes and orientations also have a contribution from weak gravitational lensing, the measurement of which is a pillar of modern observational cosmology [e.g. @kids_2020; @des_keypaper_2022; @hsc_2023]. IA can thus become an important systematic effect to weak lensing measurements, and it must be properly understood and mitigated to ensure accurate cosmological results [e.g. @Krause_2016; @samuroff_2017; @secco_2021].
+According to the halo model, galaxies form within dark matter halos, and the intrinsic shapes and orientations of these galaxies can be related with those of the host halo and with the large-scale structure of the universe (e.g. the local gravitational tidal field). This effect is known as intrinsic alignment (IA) [see, e.g., @Hirata_2004; @Blazek_2019]. The observed shapes and orientations also have a contribution from weak gravitational lensing, the measurement of which is a pillar of modern observational cosmology [e.g. @kids_2020; @des_keypaper_2022; @hsc_2023]. IA can thus become an important systematic effect on weak lensing measurements, and it must be properly understood and mitigated to ensure accurate cosmological results [e.g. @Krause_2016; @samuroff_2017; @secco_2021].
 
-Measurements of weak lensing shear help researchers study the distribution of matter and dark energy. The large-scale structure of the universe can influence the intrinsic shapes and orientations of galaxies through gravitational interactions. As such, accurately modeling this effect is important for precision cosmology with weak lensing. With upcoming surveys like the Rubin Observatory Legacy Survey of Space and Time (LSST) [@Ivezic_2019], analyses of the data will need to consider contributions from IA. A fast and flexible simulation method that includes IA is required to to provide realistic mock galaxy catalogs and to test other IA models.
[…]
