### 2025-01-21 74f67a7
Replace shortened journal tags with full journal names
