+++ paper.bib
@@ -32,7 +32,7 @@ abstract = {We present the first stable release of Halotools (v0.2), a community
 	Author = {{Hirata}, C.~M. and {Seljak}, U.},
 	Doi = {10.1103/PhysRevD.70.063526},
 	Eprint = {astro-ph/0406275},
-	Journal = {\prd},
+	Journal = {Physical Review D},
 	Keywords = {Observational cosmology, Galactic halos, Gravitational lenses and luminous arcs},
 	Month = sep,
 	Number = 6,
@@ -46,7 +46,7 @@ abstract = {We present the first stable release of Halotools (v0.2), a community
 @article{Blazek_2019,
   title = {Beyond linear galaxy alignments},
   author = {Blazek, Jonathan A. and MacCrann, Niall and Troxel, M. A. and Fang, Xiao},
-  journal = {Phys. Rev. D},
+  journal = {Physical Review D},
   volume = {100},
   issue = {10},
   pages = {103506},
@@ -61,7 +61,7 @@ abstract = {We present the first stable release of Halotools (v0.2), a community
 @article{hsc_2023,
   title = {Hyper Suprime-Cam Year 3 results: Cosmology from cosmic shear two-point correlation functions},
   author = {Li, Xiangchong and Zhang, Tianqing and Sugiyama, Sunao and Dalal, Roohi and Terasawa, Ryo and Rau, Markus M. and Mandelbaum, Rachel and Takada, Masahiro and More, Surhud and Strauss, Michael A. and Miyatake, Hironao and Shirasaki, Masato and Hamana, Takashi and Oguri, Masamune and Luo, Wentao and Nishizawa, Atsushi J. and Takahashi, Ryuichi and Nicola, Andrina and Osato, Ken and Kannawadi, Arun and Sunayama, Tomomi and Armstrong, Robert and Bosch, James and Komiyama, Yutaka and Lupton, Robert H. and Lust, Nate B. and MacArthur, Lauren A. and Miyazaki, Satoshi and Murayama, Hitoshi and Nishimichi, Takahiro and Okura, Yuki and Price, Paul A. and Tait, Philip J. and Tanaka, Masayuki and Wang, Shiang-Yu},
-  journal = {Phys. Rev. D},
+  journal = {Physical Review D},
   volume = {108},
   issue = {12},
   pages = {123518},
@@ -77,7 +77,7 @@ abstract = {We present the first stable release of Halotools (v0.2), a community
   title = {Dark Energy Survey Year 3 results: Cosmological constraints from galaxy clustering and weak lensing},
[…]
