### 2025-01-21 #7421 @nvanalfen
@ivastar Sorry for that and thank you for the clear list. Here's my summary of changes:

- Trimmed the BIB file to only references used (I tried to be very careful here, but please let me know if I have messed anything up in the rendered file).
- Replaced shortened journal references with the full journal name.
- I believe all remaining references are from the full journal, not arXiv, but please let me know if you see anything.
- I have made all (but one) of the suggested changes

For the one change I haven't (yet) made:

> Line 33: As such --> Therefore ; modeling this --> modeling of this

I have made the first change ("as such" -> "therefore"), but for the second (assuming I'm looking at the right spot), the current sentence is:

> “...accurately modeling this effect is important…”

The suggested change only altered "modeling this" to "modeling of this". I wasn't sure if you saw "accurately modeling this" as "accurate modeling this", so I have left it as is for now. If you feel that the full change
"...accurately modeling this..." to "...accurate modeling of this..." flows better, I will go right ahead and do that.

Thank you so much for all your time and hard work.
