### 2025-01-21 #7421 @ivastar
Doh, forgot one comment. Can you redo the Figure 1 caption to something like "Correlation functions from IllustrisTNG300.... showcasing the flexibility of the model. Reproduced from Figure 12 in Van Alfen et al. (2024)."
