### 2025-01-21 #7421 @ivastar
@nvanalfen we can proceed to the post-review checklist. Please follow the list above. For the release, remind me what were we going to do here? Tag a release of `halotools`? Or what's the plan?
