### 2025-01-21 a0b87e1
Fix Figure 1 caption
