+++ paper.md
@@ -49,7 +49,7 @@ Understanding and measuring IA also provides a window into the accurate modeling
 
 The new release of `halotools` creates the capability to construct realistically complex IA correlations, comparable to those of a hydrodynamic simulation, at a tiny fraction of the computational cost of a hydrodynamic simulation, as explained in @vanalfen_2024. This flexibility extends `halotools` to be of considerable benefit to simulation-based studies of IA. In @vanalfen_2024, the authors demonstrated the flexibility of the `halotools` package to create galaxy catalogs with IA comparable to various aspects of high-resolution cosmological simulations. Specifically, Figure \ref{IllustrisComparison} [taken from Figure 12 in @vanalfen_2024] shows various IA correlation functions from both IllustrisTNG300-1 [@nelson_2017; @pillepich_2017; @naiman_2017; @springel_2017; @marinacci_2017] and a galaxy catalog generated using `halotools` with its available Bolshoi-Planck (Bolplanck) halo catalog [@Klypin_2011].
 
-![Figure 12 from @vanalfen_2024 showing correlation functions from IllustrisTNG300 (points with error bars) and correlation functions measured on an HOD made with `halotools` (solid lines with shaded error regions). The parameters for the HOD model were adjusted such that the resulting correlations would match those of IllustrisTNG300, showcasing the flexibility of the model.\label{IllustrisComparison}](figures/Illustris.pdf)
+![Correlation functions from IllustrisTNG300 (points with error bars) and correlation functions measured on an HOD made with `halotools` (solid lines with shaded error regions) showcasing the flexibility of the model. REproduced from Figure 12 in @vanalfen_2024.\label{IllustrisComparison}](figures/Illustris.pdf)
 
 This release is part of a suite of modeling tools and analysis pipelines being developed to aid upcoming cosmological surveys, including LSST, Euclid, and Roman. The specific advantage of the type of models generated by `halotools` is that they are faster and lighter-weight than more expensive simulations, allowing users to quickly generate and populate catalogs of galaxies following a set of parameters. The efficiency of `halotools` also allows for direct simulation-based modeling.
 
