### 2025-01-21 #7421 @nvanalfen
@ivastar Andrew has tagged and released the IA version of halotools as v0.9.0 (it looks like minor tweaks since then have advanced to the current version being v0.9.2, tho v0.9.0 is where this work was merged in).

Should I formally mark that somewhere here?

I have also fixed and pushed the caption 1 edit. Thank you for catching that.

Also, looking at the generated PDF and I see that my author affiliation list is less of a list and more of a paragraph. It seems I haven't mastered the art of formatting that correctly. I have included them in the markdown as follows. Should I altar that?

> authors:
    - name: Nicholas Van Alfen
      orcid: 0000-0003-0049-2861
      corresponding: true
      equal-contrib: true
      affiliation: 1
    - name: Duncan Campbell
      orcid: 0000-0002-0650-9903
      equal-contrib: true
      affiliation: 2
    - name: Andrew Hearin
      orcid: 0000-0003-2219-6852
      equal-contrib: true
      affiliation: 3
    - name: Jonathan Blazek
      orcid: 0000-0002-4687-4657
      equil-contrib: true
      affiliation: 1
