### 2025-01-21 #7421 @nvanalfen
@aphearin I'm unfamiliar with the release and archive process, so I'm tagging you here in case you know the answers to what I'm about to ask. It looks like the latest version (v0.9.2) is where we included the license change. @ivastar , this seems to be the latest change we made to the code.

Is this sufficient? If so, does the PyPI entry for Halotools v0.9.2 count as the archive? I can't seem to find the DOI on that page. If not, I can (with Andrew's help potentially since I don't want to go archiving a package I'm not the maintainer for) go ahead and archive that on something more appropriate.

Apologies for all the questions and thank you for all your help.
