### 2025-01-21 #7421 @ivastar
> Also, looking at the generated PDF and I see that my author affiliation list is less of a list and more of a paragraph. It seems I haven't mastered the art of formatting that correctly. I have included them in the markdown as follows. Should I alter that?
>
> >

This is the typical formatting. I don't see anything wrong with it. It's ok to leave as is.
