### 2025-01-21 #7421 @ivastar
> does the PyPI entry for Halotools v0.9.2 count as the archive?

The PyPI release does not count as archive. It should be possible to archive this version to Zenodo, seems like there are 2 releases already: https://zenodo.org/records/835895
