### 2025-01-27 #7421 @ivastar
@nvanalfen @aphearin quoting from the acting editor in chief:

"The principle behind the archive, I believe, is to have a permanent archive of the code that was reviewed for the paper, which to me implies that we should archive the fork, unless the reviewed code is merged into the upstream `astropy/halotools` repo? In that case, just pointing to the appropriate version in the existing `astropy/halotools` Zenodo record seems fine to me."

According to the pre-review discussion, the changes here have already been merged upstrea. So, does any of the existing Zenodo archives have these changes? In that case, we can point to that DOI. Or can you make a new Zenodo archive of `halotools` which includes the changes?
