### 2025-01-27 #7421 @ivastar
> All of the code changes have been merged into the official `astropy/halotools`, so the only differences between the repos right now have to do with the paper. It looks like I'll have to make a new Zenodo record for the most recent version of `halotools`. I'll work with Andrew on that.
>
> [@ivastar](https://github.com/ivastar) Two questions:
>
> * There is a point in that list you sent me that says to make sure the author list on the archive matches the author list in the JOSS paper. The author list in this paper only includes those who worked on the IA functionality, where the author list for the archive of the latest version of halotools should include everyone who had made contributions to the code beyond the IA parts (other bits have been developed and merged in alongside what I have been doing for this IA-specific release). Is this still fine?

Yes, that's ok in this case as long as all authors of the paper are included as archive authors.

> * I am a little confused about the requirement that says the titles must match. The paper of course says "Halotools: a new release..." but the archive would be the simpler "astropy/halotools: v0.9.2" to simply note the version it was released as, correct?

Also, ok to have a mismatch in this case.
