### 2025-01-28 #7421 @nvanalfen
@ivastar Sorry for the delay. Just and update: Andrew and I are working on getting access to the existing Zenodo archive for halotools so I can upload a new version. Once I manage that, I will post the DOI and other required information here. Thank you for your patience and help.
