### 2025-02-25 #7421 @nvanalfen
@ivastar Sorry about the delay. We were trying to get permission to update the existing halotools Zenodo entry with a new version. We weren't able to, but Andrew managed to make a new entry (and added a soft reference to the old one under linked works).

Important and checklist:
- Andrew made a release for 0.9.3 (one later than what I had listed above as 0.9.2) to ensure compatibility with numpy 2. This will be the version we will use.
- Andrew has uploaded the Zenodo archive for this version (0.9.3)
- The doi for this version is https://doi.org/10.5281/zenodo.14920341
- Andrew, Duncan, Jonathan, and I are all listed as authors on Zenodo (Andrew as creator and the other three of us as contributors)
- ORCIDs are all correct for the four of us
- The license on Zenodo is the BSD 3-clause license. The same as on GitHub.
- The Zenodo title doesn't match the paper, but we discussed earlier and you mentioned it was fine to have a mismatch in this case.

Let me know what else is needed on my end. Thank you so much for your patience and help.
