### 2025-03-04 #7421 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
