### 2025-03-05 e612ed2
editorial tweaks ahead of publication in JOSS
