+++ paper.md
@@ -1,12 +1,13 @@
 ---
-title: 'Halotools: A New Release Adding Intrinsic Alignments to Halo Based Methods'
+title: 'Halotools: A New Release Adding Intrinsic Alignments to Halo-Based Methods'
 tags:
     - Python
-    - Cosmology
-    - Weak Gravitational Lensing
-    - Intrinsic Alignments
+    - cosmology
+    - weak gravitational lensing
+    - intrinsic alignments
 authors:
-    - name: Nicholas Van Alfen
+    - given-names: Nicholas
+	  surname: Van Alfen
       orcid: 0000-0003-0049-2861
       corresponding: true
       equal-contrib: true
@@ -35,36 +36,36 @@ bibliography: paper.bib
 ---
 
 # Summary
-`Halotools`, originally published in 2017, is a Python package for cosmology and astrophysics designed to generate mock universes using existing catalogs of dark matter halos [@Hearin_2017]. A theoretical basis of the library is the so-called halo model, that describes the matter distribution of dark matter as gravitationally self-bound clouds of dark matter particles that we call halos. `Halotools` is designed to take an underlying catalog of dark matter halos and populate them with galaxies using subhalo abundance, or halo occupation distribution (HOD) models, creating catalogs of simulated galaxies for use in research. This release (v0.9) adds functionality to align galaxies, injecting what are known as intrinsic alignments (IA) into these catalogs. As a result, these simulated galaxy catalogs can now be created with realistically complex correlations between galaxies, mimicking some effects seen in more expensive hydrodynamic simulations.
+Halotools, originally published in 2017, is a Python package for cosmology and astrophysics designed to generate mock universes using existing catalogs of dark matter halos [@Hearin_2017]. A theoretical basis of the library is the so-called halo model, which describes the matter distribution of dark matter as gravitationally self-bound clouds of dark matter particles that we call halos. Halotools is designed to take an underlying catalog of dark matter halos and populate them with galaxies using subhalo abundance, or halo occupation distribution (HOD) models, creating catalogs of simulated galaxies for use in research. This release (v0.9) adds functionality to align galaxies, injecting what are known as intrinsic alignments (IA) into these catalogs. As a result, these simulated galaxy catalogs can now be created with realistically complex correlations between galaxies, mimicking some effects seen in more expensive hydrodynamic simulations.
 
 # Statement of Need
-According to the halo model, galaxies form within dark matter halos, and the intrinsic shapes and orientations of these galaxies can be related with those of the host halo and with the large-scale structure of the universe (e.g. the local gravitational tidal field). This effect is known as intrinsic alignment (IA) [see, e.g., @Hirata_2004; @Blazek_2019]. The observed shapes and orientations also have a contribution from weak gravitational lensing, the measurement of which is a pillar of modern observational cosmology [e.g. @kids_2020; @des_keypaper_2022; @hsc_2023]. IA can thus become an important systematic effect on weak lensing measurements, and it must be properly understood and mitigated to ensure accurate cosmological results [e.g. @Krause_2016; @samuroff_2017; @secco_2021].
+According to the halo model, galaxies form within dark matter halos, and the intrinsic shapes and orientations of these galaxies can be related with those of the host halo and with the large-scale structure of the universe (e.g. the local gravitational tidal field). This effect is known as intrinsic alignment [IA, see, e.g., @Blazek_2019; @Hirata_2004]. The observed shapes and orientations also have a contribution from weak gravitational lensing, the measurement of which is a pillar of modern observational cosmology [e.g. @des_keypaper_2022; @kids_2020; @hsc_2023]. IA can thus become an important systematic effect on weak lensing measurements, and it must be properly understood and mitigated to ensure accurate cosmological results [e.g. @Krause_2016; @samuroff_2017; @secco_2021].
 
[…]
