### 2025-03-05 #7421 @warrickball
Hi! I've opened a [small PR](https://github.com/nvanalfen/halotools/pull/10) with just a handful of tweaks. Let me know when you've checked these over and merged the PR and I'll give the publication pipeline another test run. I'm hoping that's it (there was very little left to fix!) and we'll be able to publish soon.
