+++ paper.md
@@ -43,7 +43,7 @@ According to the halo model, galaxies form within dark matter halos, and the int
 
 Measurements of weak lensing shear help researchers study the distribution of matter and dark energy. The large-scale structure of the Universe can influence the intrinsic shapes and orientations of galaxies through gravitational interactions. Therefore, accurately modeling this effect is important for precision cosmology with weak lensing. With upcoming surveys such as the Rubin Observatory Legacy Survey of Space and Time [LSST, @Ivezic_2019], analyses of the data will need to consider contributions from IA. A fast and flexible simulation method that includes IA is required to to provide realistic mock galaxy catalogs and to test other IA models.
 
-Understanding and measuring IA also provides a window into the accurate modeling of galaxy formation and a is probe of cosmic structure and potentially new physics [e.g. @chisari_2013]. Halotools already provides tools for modeling the relationship between galaxies and the halos in which they reside (the galaxy--halo connection), and it is widely used in the field. The expanded functionality added in this release allows Halotools to be used to produce mock galaxy catalogs with realistically complex galaxy orientations. These catalogs can then be used to test and validate IA models, to study IA in observational data and in hydrodynamic simulations [e.g. @nelson_2017; @pillepich_2017; @naiman_2017; @springel_2017; @marinacci_2017], and to provide a fully nonlinear, simulation-based model for observed galaxy clustering and lensing statistics.
+Understanding and measuring IA also provides a window into the accurate modeling of galaxy formation and is a probe of cosmic structure and potentially new physics [e.g. @chisari_2013]. Halotools already provides tools for modeling the relationship between galaxies and the halos in which they reside (the galaxy--halo connection), and it is widely used in the field. The expanded functionality added in this release allows Halotools to be used to produce mock galaxy catalogs with realistically complex galaxy orientations. These catalogs can then be used to test and validate IA models, to study IA in observational data and in hydrodynamic simulations [e.g. @nelson_2017; @pillepich_2017; @naiman_2017; @springel_2017; @marinacci_2017], and to provide a fully nonlinear, simulation-based model for observed galaxy clustering and lensing statistics.
 
 # Significance
 Halotools provides a way for users to create halo occupation models such as abundance matching and the halo occupation distribution (HOD), and enables a modular approach to mock universe creation. The user can provide a series of component models to the HOD model describing features that will govern how Halotools populates these dark matter halos with galaxies. This release provides methods to describe galaxy alignment, including IA similarity, analogous to how other features in HOD models are defined.
@@ -55,7 +55,7 @@ The new release of Halotools creates the capability to construct realistically c
 This release is part of a suite of modeling tools and analysis pipelines being developed to aid upcoming cosmological surveys, including LSST, Euclid, and Roman. The specific advantage of the type of models generated by Halotools is that they are faster and lighter-weight than more expensive simulations, allowing users to quickly generate and populate catalogs of galaxies following a set of parameters. The efficiency of Halotools also allows for direct simulation-based modeling.
 
 # Structure
-Currently, the user needs to provide the following in order to build a mock galaxy catalog with IA using Halotools (optional components are in parentheses):
+Currently, the user needs to provide the following in order to build a mock galaxy catalog with IA using Halotools:
 
 \begin{itemize}
     \item Occupation Model: Determines the number density of galaxies within a given halo.
