### 2025-03-10 #7421 @warrickball
Thanks @nvanalfen! I opened [another PR](https://github.com/nvanalfen/halotools/pull/11) for two more very small errors that should truly be the last thing before we can publish.
