### 2025-03-11 0dca18a
Make stub for potential flexible custom alignment strength model
