### 2025-03-11 b646106
Make custom alignment strength model classes for both central and satellite galaxies
