### 2025-03-11 89e3786
Move common functionality into _CustomAlignmentStrengthTemplate and ansure alignment strengths clip between -1 and 1
