### 2025-03-11 ea6194a
Expand author list in LSST reference
