+++ paper.bib
@@ -147,20 +147,22 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Ivezic_2019,
-author = {{Ivezi{\'c}}, {\v Z}. and {Kahn}, S.~M. and {Tyson}, J.~A. and {Abel}, B. and {Acosta}, E. and {Allsman}, R. and {Alonso}, D. and {AlSayyad}, Y. and {Anderson}, S.~F. and {Andrew}, J. and et al.},
-title = {LSST: From Science Drivers to Reference Design and Anticipated Data Products},
-journal = {The Astrophysical Journal},
-archivePrefix = "arXiv",
-eprint = {0805.2366},
-keywords = {astrometry, cosmology: observations, Galaxy: general, methods: observational, stars: general, surveys}, 
-year = 2019,
-month = mar,
-volume = 873,
-eid = {111},
-pages = {111},
-doi = {10.3847/1538-4357/ab042c},
-adsurl = {http://adsabs.harvard.edu/abs/2019ApJ...873..111I},
-adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+       author = {{Ivezi{\'c}}, {\v{Z}}eljko and {Kahn}, Steven M. and {Tyson}, J. Anthony and {Abel}, Bob and {Acosta}, Emily and {Allsman}, Robyn and {Alonso}, David and {AlSayyad}, Yusra and {Anderson}, Scott F. and {Andrew}, John and {Angel}, James Roger P. and {Angeli}, George Z. and {Ansari}, Reza and {Antilogus}, Pierre and {Araujo}, Constanza and {Armstrong}, Robert and {Arndt}, Kirk T. and {Astier}, Pierre and {Aubourg}, {\'E}ric and {Auza}, Nicole and {Axelrod}, Tim S. and {Bard}, Deborah J. and {Barr}, Jeff D. and {Barrau}, Aurelian and {Bartlett}, James G. and {Bauer}, Amanda E. and {Bauman}, Brian J. and {Baumont}, Sylvain and {Bechtol}, Ellen and {Bechtol}, Keith and {Becker}, Andrew C. and {Becla}, Jacek and {Beldica}, Cristina and {Bellavia}, Steve and {Bianco}, Federica B. and {Biswas}, Rahul and {Blanc}, Guillaume and {Blazek}, Jonathan and {Blandford}, Roger D. and {Bloom}, Josh S. and {Bogart}, Joanne and {Bond}, Tim W. and {Booth}, Michael T. and {Borgland}, Anders W. and {Borne}, Kirk and {Bosch}, James F. and {Boutigny}, Dominique and {Brackett}, Craig A. and {Bradshaw}, Andrew and {Brandt}, William Nielsen and {Brown}, Michael E. and {Bullock}, James S. and {Burchat}, Patricia and {Burke}, David L. and {Cagnoli}, Gianpietro and {Calabrese}, Daniel and {Callahan}, Shawn and {Callen}, Alice L. and {Carlin}, Jeffrey L. and {Carlson}, Erin L. and {Chandrasekharan}, Srinivasan and {Charles-Emerson}, Glenaver and {Chesley}, Steve and {Cheu}, Elliott C. and {Chiang}, Hsin-Fang and {Chiang}, James and {Chirino}, Carol and {Chow}, Derek and {Ciardi}, David R. and {Claver}, Charles F. and {Cohen-Tanugi}, Johann and {Cockrum}, Joseph J. and {Coles}, Rebecca and {Connolly}, Andrew J. and {Cook}, Kem H. and {Cooray}, Asantha and {Covey}, Kevin R. and {Cribbs}, Chris and {Cui}, Wei and {Cutri}, Roc and {Daly}, Philip N. and {Daniel}, Scott F. and {D
[…]
