### 2025-03-11 55fa34b
set title of LSST reference in sentence case
