+++ paper.bib
@@ -148,7 +148,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Ivezic_2019,
        author = {{Ivezi{\'c}}, {\v{Z}}eljko and {Kahn}, Steven M. and {Tyson}, J. Anthony and {Abel}, Bob and {Acosta}, Emily and {Allsman}, Robyn and {Alonso}, David and {AlSayyad}, Yusra and {Anderson}, Scott F. and {Andrew}, John and {Angel}, James Roger P. and {Angeli}, George Z. and {Ansari}, Reza and {Antilogus}, Pierre and {Araujo}, Constanza and {Armstrong}, Robert and {Arndt}, Kirk T. and {Astier}, Pierre and {Aubourg}, {\'E}ric and {Auza}, Nicole and {Axelrod}, Tim S. and {Bard}, Deborah J. and {Barr}, Jeff D. and {Barrau}, Aurelian and {Bartlett}, James G. and {Bauer}, Amanda E. and {Bauman}, Brian J. and {Baumont}, Sylvain and {Bechtol}, Ellen and {Bechtol}, Keith and {Becker}, Andrew C. and {Becla}, Jacek and {Beldica}, Cristina and {Bellavia}, Steve and {Bianco}, Federica B. and {Biswas}, Rahul and {Blanc}, Guillaume and {Blazek}, Jonathan and {Blandford}, Roger D. and {Bloom}, Josh S. and {Bogart}, Joanne and {Bond}, Tim W. and {Booth}, Michael T. and {Borgland}, Anders W. and {Borne}, Kirk and {Bosch}, James F. and {Boutigny}, Dominique and {Brackett}, Craig A. and {Bradshaw}, Andrew and {Brandt}, William Nielsen and {Brown}, Michael E. and {Bullock}, James S. and {Burchat}, Patricia and {Burke}, David L. and {Cagnoli}, Gianpietro and {Calabrese}, Daniel and {Callahan}, Shawn and {Callen}, Alice L. and {Carlin}, Jeffrey L. and {Carlson}, Erin L. and {Chandrasekharan}, Srinivasan and {Charles-Emerson}, Glenaver and {Chesley}, Steve and {Cheu}, Elliott C. and {Chiang}, Hsin-Fang and {Chiang}, James and {Chirino}, Carol and {Chow}, Derek and {Ciardi}, David R. and {Claver}, Charles F. and {Cohen-Tanugi}, Johann and {Cockrum}, Joseph J. and {Coles}, Rebecca and {Connolly}, Andrew J. and {Cook}, Kem H. and {Cooray}, Asantha and {Covey}, Kevin R. and {Cribbs}, Chris and {Cui}, Wei and {Cutri}, Roc and {Daly}, Philip N. and {Daniel}, Scott F. and {Daruich}, Felipe and {Daubard}, Guillaume and {Daues}, Greg and {Dawson}, William and {Delgado}, Francisco and {Dellapenna}, Alfred and {de Peyster}, Robert and {de Val-Borro}, Miguel and {Digel}, Seth W. and {Doherty}, Peter and {Dubois}, Richard and {Dubois-Felsmann}, Gregory P. and {Durech}, Josef and {Economou}, Frossie and {Eifler}, Tim and {Eracleous}, Michael and {Emmons}, Benjamin L. and {Fausti Neto}, Angelo and {Ferguson}, Henry and {Figueroa}, Enrique and {Fisher-Levine}, Merlin and {Focke}, Warren and {Foss}, Michael D. and {Frank}, James and {Freemon}, Michael D. and {Gangler}, Emmanuel and {Gawiser}, Eric and {Geary}, John C. and {Gee}, Perry and {Geha}, Marla and {Gessner}, Charles J.~B. and {Gibson}, Robert
[…]
