### 2025-03-11 #7421 @warrickball
I apologise: [one more](https://github.com/nvanalfen/halotools/pull/12)!
