### 2025-03-11 #7421 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
