### 2025-03-11 #7421 @warrickball
Thanks once again to @matroxel, @cmlamman and @fjcastander for reviewing and @ivastar for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.

Congratulations @nvanalfen, your paper has been published in JOSS! 🎉
